requirements.txt
requests==2.31.0
Flask==3.0.0
oauthlib==3.2.2
pyOpenSSL==23.3.0
Flask-Login==0.6.3
python-dotenv==1.0.0
pymysql==1.1.0
flask-wtf==1.2.1